Added base command object and moved parser to package
